pyproject.toml
[project]
name = "cadastro"
version = "0.1.0"
description = "cadastro"
authors = [
    {name = "joao", email = "joao.costa.graciano@gmail.com"},
]
dependencies = ["django>=4.2.18", "djangorestframework>=3.15.2"]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
